move logging from core to bin, and more
also:
- prepare for changing log_level at runtime
- refactor `setup_logging` into separta functions depending on
  build-flag
