+++ leftwm-core/src/handlers/command_handler.rs
@@ -83,7 +83,7 @@ fn process_internal<C: Config, SERVER: DisplayServer>(
         Command::SendWorkspaceToTag(ws_index, tag_index) => {
             Some(send_workspace_to_tag(manager, *ws_index, *tag_index))
         }
-        Command::SetLogLevel(loglevel) => set_log_level(loglevel),
+        Command::SetLogLevel(loglevel) => set_log_level(manager, loglevel),
         Command::Other(cmd) => Some(C::command_handler(cmd, manager)),
     }
 }
@@ -612,8 +612,8 @@ fn set_log_level<C: Config, SERVER: DisplayServer>(
     manager: &mut Manager<C, SERVER>,
     loglevel: &String,
 ) -> Option<bool> {
-    manager.state.config.loglevel = logleve;
-    true
+    manager.state.loglevel = loglevel.to_string();
+    return Some(true);
 }
 
 #[cfg(test)]
