+++ leftwm-core/src/lib.rs
@@ -24,7 +24,6 @@ pub mod errors;
 mod event_loop;
 mod handlers;
 pub mod layouts;
-pub mod logging;
 pub mod models;
 mod state;
 pub mod utils;
